[#266] Move encryption to a concern
